Create prototype of events module
The module exports functions that match types of events in the log
(currently only state_goal is stubbed). Main module is calling
appropriate function for each event in a log.

The goal is to have other types of events and to create an in memory
representation of the state (days, goals, opportunities, relations) in
an events sourcing manner.